Move tests for API resources to a different package
To prevent the API application to load the test files which can cause class not found exception
