+++ pinot-controller/src/test/java/org/apache/pinot/controller/api/AccessControlTest.java
@@ -16,10 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.controller.api.resources;
+package org.apache.pinot.controller.api;
 
-import java.io.File;
+import java.io.IOException;
 import org.apache.pinot.controller.ControllerConf;
+import org.apache.pinot.controller.api.access.AccessControl;
+import org.apache.pinot.controller.api.access.AccessControlFactory;
 import org.apache.pinot.controller.helix.ControllerTest;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -27,32 +29,39 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 
-public class FileUploadPathProviderTest extends ControllerTest {
-  private static ControllerConf _controllerConf;
+public class AccessControlTest extends ControllerTest {
 
   @BeforeClass
   public void setUp() {
     startZk();
-    _controllerConf = getDefaultControllerConfiguration();
-    startController(_controllerConf);
+    ControllerConf config = getDefaultControllerConfiguration();
+    config.setAccessControlFactoryClass(DenyAllAccessFactory.class.getName());
+    startController(config);
   }
 
   @Test
-  public void testFileUploadPathProvider()
-      throws Exception {
-    FileUploadPathProvider provider = new FileUploadPathProvider(_controllerConf);
-    Assert.assertEquals(provider.getBaseDataDir().getAbsolutePath(), _controllerDataDir);
-    String fileUploadTmpDirPath = provider.getFileUploadTmpDir().getAbsolutePath();
-    Assert.assertEquals(fileUploadTmpDirPath, new File(_controllerDataDir, "fileUploadTemp").getAbsolutePath());
-    Assert.assertEquals(provider.getSchemasTmpDir().getAbsolutePath(),
-        new File(_controllerDataDir, "schemasTemp").getAbsolutePath());
-    Assert.assertEquals(provider.getTmpUntarredPath().getAbsolutePath(),
-        new File(fileUploadTmpDirPath, "untarred").getAbsolutePath());
+  public void testAccessDenied() {
+    try {
+      sendGetRequest(_controllerRequestURLBuilder.forSegmentDownload("testTable", "testSegment"));
+    } catch (IOException e) {
+      Assert.assertTrue(e.getMessage().startsWith("Server returned HTTP response code: 403"));
+      return;
+    }
+    Assert.fail("Access not denied");
   }
 
   @AfterClass
   public void tearDown() {
     stopController();
     stopZk();
+  }
+
+  public static class DenyAllAccessFactory implements AccessControlFactory {
+    private static final AccessControl DENY_ALL_ACCESS = (httpHeaders, tableName) -> false;
+
+    @Override
+    public AccessControl create() {
+      return DENY_ALL_ACCESS;
+    }
   }
 }
